only node 0 should mess with files; workaround ga header problem

+++ src/python/nw_inp_from_string.c
@@ -1,3 +1,4 @@
+#include "global.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
@@ -14,31 +15,29 @@ extern int string_to_fortchar(char *, int, char *);
 int nw_inp_from_string(int rtdb, const char *input)
 {
     char *filename = "temp.nw";
-    FILE *file = fopen(filename,"w");
+    FILE *file;
     char fstring[255];
     int status;
 
-    if (!file) {
-	fprintf(stderr,"failed to open temp.nw\n");
-	return 0;
-    }
+    if (ga_nodeid_() == 0) {
+      if (!(file = fopen(filename,"w"))) {
+	ga_error("nw_inp_from_string: failed to open temp.nw\n",0);
+      }
 
-    if (fwrite(input, 1, strlen(input), file) != strlen(input)) {
-	fprintf(stderr,"failed to write to temp.nw\n");
-	(void) fclose(file);
-	return 0;
-    }
+      if (fwrite(input, 1, strlen(input), file) != strlen(input)) {
+	ga_error("nw_inp_from_string: failed to write to temp.nw\n",0);
+      }
 
-    (void) fclose(file);
+      (void) fclose(file);
+    }
 
     if (!string_to_fortchar(fstring, sizeof(fstring), filename)) {
-	fprintf(stderr,"fstring is too small?\n");
-	return 0;
+      ga_error("nw_inp_from_string: fstring is too small",0);
     }
 
     status = NWC_NW_INP_FROM_FILE(&rtdb, fstring, sizeof fstring);
 
-    (void) unlink(filename);
+    if (ga_nodeid_() == 0) (void) unlink(filename);
 
     return status;
 }
